Refactor kevent() to make the flow of control more obvious.
Add reference counting to the knote structure. This is slightly buggy and
probably unnecessary.

Eliminate the kqueue_lock() and kqueue_unlock() functions, and create 
a heirarchy of two locks: the knotelist lock at the filter level, and
a per-knote lock. This needs to be looked at more closely.

Make note of a race condition when kev.flags contains EV_ADD | EV_DISABLE
and a new knote is created.



git-svn-id: svn://mark.heily.com/libkqueue/trunk@411 34a59efb-09c4-4dcf-b95f-994a32aba0d8